Use recipe that handles restarting the server
Can't just enable the management plugin.  The server needs to be
restarted as well after the plugin is enabled before the plugin starts
running (or at least the Web server portion of it)  There's a recipe
that seems to do that.  We'll delegate to it instead of directly
enabling the plugin